+++ Menu/Objects/ButtonScrollerObjects.cs
@@ -5,16 +5,13 @@ using System.Text;
 using System.Threading.Tasks;
 using Menu;
 using Menu.Remix.MixedUI;
+using RainMeadow.UI.Interfaces;
 using UnityEngine;
-using static RainMeadow.ButtonScroller;
 
 namespace RainMeadow.UI.Components
 {
-    public class ScrollSymbolButton : SimplerSymbolButton, IPartOfButtonScroller
+    public class ScrollSymbolButton : SimplerSymbolButton, IOwnMenuScrollObject
     {
-        public float Alpha { get; set; } = 1;
-        public Vector2 Pos { get => pos; set => pos = value; }
-        public Vector2 Size { get => size; set => size = value; }
         public ScrollSymbolButton(Menu.Menu menu, MenuObject owner, string symbolName, string signalText, Vector2 pos, Vector2 size = default) : base(menu, owner, symbolName, signalText, pos)
         {
             this.size = size == default ? this.size : size;
@@ -26,7 +23,7 @@ namespace RainMeadow.UI.Components
         }
 
     }
-    public class AlignedMenuLabel(Menu.Menu menu, MenuObject owner, string text, Vector2 pos, Vector2 size, bool bigText, FTextParams txtParams = null) : MenuLabel(menu, owner, text, pos, size, bigText, txtParams), ButtonScroller.IPartOfButtonScroller
+    public class AlignedMenuLabel(Menu.Menu menu, MenuObject owner, string text, Vector2 pos, Vector2 size, bool bigText, FTextParams txtParams = null) : MenuLabel(menu, owner, text, pos, size, bigText, txtParams), IOwnMenuScrollObject
     {
         public float Alpha { get; set; } = 1;
         public Vector2 Pos { get => pos; set => pos = value; }