+++ OnlineUIComponents/ChatInputOverlay.cs
@@ -10,6 +10,7 @@ namespace RainMeadow
 
         public ChatInputOverlay(ProcessManager manager) : base(RMOverlayHUDMenu.GetOverlayMenu(), RMOverlayHUDMenu.GetOverlayMenu().pages[0])
         {
+            RainMeadow.DebugMe();
             chat = new ChatTextBox(this.menu, this, "", new Vector2(manager.rainWorld.options.ScreenSize.x * 0.001f + (1366f - manager.rainWorld.options.ScreenSize.x) / 2f, 0), new(750, 30));
             this.subObjects.Add(chat);
         }